Split forge tests, so they can be filtered out
