+++ starknet-foundry/crates/forge/tests/integration/declare.rs
@@ -1,5 +1,5 @@
-use crate::common::corelib::{corelib, predeployed_contracts};
-use crate::common::runner::Contract;
+use crate::integration::common::corelib::{corelib, predeployed_contracts};
+use crate::integration::common::runner::Contract;
 use crate::{assert_failed, assert_passed, test_case};
 use camino::Utf8PathBuf;
 use forge::run;
