+++ starknet-foundry/crates/forge/tests/main.rs
@@ -1,0 +1,2 @@
+mod e2e;
+mod integration;
